[#67] Episodes progress via web-socket: creating common WS view
