+++ src/common/schemas.py
@@ -1,0 +1,9 @@
+from marshmallow import fields, Schema
+
+
+class WSHeadersRequestSchema(Schema):
+    authorization = fields.Str()
+
+
+class WSRequestAuthSchema(Schema):
+    headers = fields.Nested(WSHeadersRequestSchema())
